+++ packages/cli/__tests__/zosfiles/__system__/search/ds/cli.files.search.ds.system.test.ts
@@ -36,7 +36,7 @@ let pdsBadMemNames: string[];
 let searchString = "Zowe CLI";
 let oldForceColor: string;
 
-const goodTestString = "This system test is brought to you by Zowe CLI!";
+const goodTestString = "^ This system test is brought to you by Zowe CLI!";
 const badTestString = "Sadly, this string will not match the search.";
 
 describe("Search Data Sets", () => {
@@ -66,9 +66,9 @@ describe("Search Data Sets", () => {
         for (const dsn of [...goodDsNames, ...badDsNames]) {
             await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_SEQUENTIAL, dsn);
             if (goodDsNames.includes(dsn)) {
-                await Upload.bufferToDataSet(REAL_SESSION, Buffer.from(goodTestString), `${dsn}`);
+                await Upload.bufferToDataSet(REAL_SESSION, Buffer.from(goodTestString), `${dsn}`, {encoding: "IBM-1047"});
             } else {
-                await Upload.bufferToDataSet(REAL_SESSION, Buffer.from(badTestString), `${dsn}`);
+                await Upload.bufferToDataSet(REAL_SESSION, Buffer.from(badTestString), `${dsn}`, {encoding: "IBM-1047"});
             }
         }
 
@@ -133,7 +133,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should search and find the correct data sets rfj", async () => {
@@ -147,13 +147,13 @@ describe("Search Data Sets", () => {
                 "--rfj"
             ]);
             const expectedApiResponse = [
-                {dsn: `${dsnPrefix}.PDS1`, member: "MEM2", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.PDS1`, member: "MEM3", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.PDS2`, member: "MEM2", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.PDS2`, member: "MEM3", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.SEQ1`, matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.SEQ4`, matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.SEQ5`, matchList: [{line: 1, column: 39, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS1`, member: "MEM2", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS1`, member: "MEM3", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS2`, member: "MEM2", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS2`, member: "MEM3", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.SEQ1`, matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.SEQ4`, matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.SEQ5`, matchList: [{line: 1, column: 41, contents: goodTestString}]},
             ];
 
             expect(response.stderr.toString()).toBe("");
@@ -167,7 +167,32 @@ describe("Search Data Sets", () => {
             expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
+        });
+
+        it("should search and find the correct data sets with encoding", async () => {
+            const response = runCliScript(shellScript, TEST_ENVIRONMENT_NO_PROF, [
+                pattern,
+                searchString,
+                defaultSys.zosmf.host,
+                defaultSys.zosmf.port,
+                defaultSys.zosmf.user,
+                defaultSys.zosmf.password,
+                "--encoding IBM-037"
+            ]);
+
+            expect(response.stderr.toString()).toBe("");
+            expect(response.status).toBe(0);
+            expect(response.stdout.toString()).toContain(`Found "${searchString}" in 7 data sets and PDS members`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS1" | Member "MEM2":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS1" | Member "MEM3":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS2" | Member "MEM2":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS2" | Member "MEM3":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
+            expect(response.stdout.toString()).not.toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ¬${goodTestString.substring(1)}`);
         });
 
         it("should perform an initial mainframe search if requested", async () => {
@@ -191,7 +216,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should handle case sensitive searches 1", async () => {
@@ -215,7 +240,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should handle case sensitive searches 2", async () => {
@@ -240,7 +265,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).not.toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).not.toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).not.toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).not.toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).not.toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should allow for multiple concurrent requests", async () => {
@@ -264,7 +289,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should time out after some time 1", async () => {
@@ -288,7 +313,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should time out after some time 2", async () => {
@@ -331,19 +356,36 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
+        });
+
+        it("should search and find the correct data sets with encoding - part 1", async () => {
+            const response = runCliScript(shellScript, TEST_ENVIRONMENT, [pattern, searchString, "--encoding IBM-037"]);
+
+            expect(response.stderr.toString()).toBe("");
+            expect(response.status).toBe(0);
+            expect(response.stdout.toString()).toContain(`Found "${searchString}" in 7 data sets and PDS members`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS1" | Member "MEM2":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS1" | Member "MEM3":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS2" | Member "MEM2":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.PDS2" | Member "MEM3":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
+            expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
+            expect(response.stdout.toString()).not.toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ¬${goodTestString.substring(1)}`);
         });
 
         it("should search and find the correct data sets rfj", async () => {
             const response = runCliScript(shellScript, TEST_ENVIRONMENT, [pattern, searchString, "--rfj"]);
             const expectedApiResponse = [
-                {dsn: `${dsnPrefix}.PDS1`, member: "MEM2", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.PDS1`, member: "MEM3", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.PDS2`, member: "MEM2", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.PDS2`, member: "MEM3", matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.SEQ1`, matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.SEQ4`, matchList: [{line: 1, column: 39, contents: goodTestString}]},
-                {dsn: `${dsnPrefix}.SEQ5`, matchList: [{line: 1, column: 39, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS1`, member: "MEM2", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS1`, member: "MEM3", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS2`, member: "MEM2", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.PDS2`, member: "MEM3", matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.SEQ1`, matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.SEQ4`, matchList: [{line: 1, column: 41, contents: goodTestString}]},
+                {dsn: `${dsnPrefix}.SEQ5`, matchList: [{line: 1, column: 41, contents: goodTestString}]},
             ];
 
             expect(response.stderr.toString()).toBe("");
@@ -357,7 +399,7 @@ describe("Search Data Sets", () => {
             expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(JSON.parse(response.stdout.toString()).data.commandResponse).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should perform an initial mainframe search if requested", async () => {
@@ -373,7 +415,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should handle case sensitive searches 1", async () => {
@@ -389,7 +431,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should handle case sensitive searches 2", async () => {
@@ -406,7 +448,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).not.toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).not.toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).not.toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).not.toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).not.toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should allow for multiple concurrent requests", async () => {
@@ -422,7 +464,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should time out after some time 1", async () => {
@@ -438,7 +480,7 @@ describe("Search Data Sets", () => {
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ1":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ4":`);
             expect(response.stdout.toString()).toContain(`Data Set "${dsnPrefix}.SEQ5":`);
-            expect(response.stdout.toString()).toContain(`Line: 1, Column: 39, Contents: ${goodTestString}`);
+            expect(response.stdout.toString()).toContain(`Line: 1, Column: 41, Contents: ${goodTestString}`);
         });
 
         it("should time out after some time 2", async () => {